Add files via upload
Imagine you are driving on the highway, and your car suddenly breaks down or you are end up with fuel. You're miles away from the nearest town and you have no help service. What do you do? On Road Vehicle Breakdown Help Assistance is an Android application that will help you in such a situation.
this project is Help assistance. 

this is a firebase project. it's was a group project.